Restructuring for example_1 (#234)
* First compilable version

* Fix wrong paths

* Fix rst

* Update copyright


* Restructure cmakelists

* Add maintainer

* Cleanup old paths

* Update Dr. Denis' Email and add him as author.

* Update maintainer

* Use test_nodes from ros2_controllers

* Fix trailing whitespaces

* Fix rviz display

* Add image to readme

---------
