+++ example_1/bringup/launch/test_forward_position_controller.launch.py
@@ -22,7 +22,7 @@ def generate_launch_description():
 
     position_goals = PathJoinSubstitution(
         [
-            FindPackageShare("ros2_control_demo_bringup"),
+            FindPackageShare("ros2_control_demo_example_1"),
             "config",
             "rrbot_forward_position_publisher.yaml",
         ]
@@ -31,7 +31,7 @@ def generate_launch_description():
     return LaunchDescription(
         [
             Node(
-                package="ros2_control_test_nodes",
+                package="ros2_controllers_test_nodes",
                 executable="publisher_forward_position_controller",
                 name="publisher_forward_position_controller",
                 parameters=[position_goals],
