+++ example_1/description/launch/view_robot.launch.py
@@ -26,7 +26,7 @@ def generate_launch_description():
     declared_arguments.append(
         DeclareLaunchArgument(
             "description_package",
-            default_value="rrbot_description",
+            default_value="ros2_control_demo_example_1",
             description="Description package with robot URDF/xacro files. Usually the argument \
         is not set, it enables use of a custom description.",
         )
@@ -69,7 +69,7 @@ def generate_launch_description():
     robot_description = {"robot_description": robot_description_content}
 
     rviz_config_file = PathJoinSubstitution(
-        [FindPackageShare(description_package), "config", "rrbot.rviz"]
+        [FindPackageShare(description_package), "rviz", "rrbot.rviz"]
     )
 
     joint_state_publisher_node = Node(
