+++ example_1/description/meshes/hokuyo.dae
