+++ example_1/description/urdf/rrbot.urdf.xacro
@@ -8,13 +8,13 @@ https://github.com/ros-simulation/gazebo_ros_demos/blob/kinetic-devel/rrbot_desc
   <xacro:arg name="prefix" default="" />
 
   <!-- Import RRBot macro -->
-  <xacro:include filename="$(find rrbot_description)/urdf/rrbot_description.urdf.xacro" />
+  <xacro:include filename="$(find ros2_control_demo_example_1)/urdf/rrbot_description.urdf.xacro" />
 
   <!-- Import Rviz colors -->
-  <xacro:include filename="$(find rrbot_description)/gazebo/rrbot.materials.xacro" />
+  <xacro:include filename="$(find ros2_control_demo_example_1)/urdf/rrbot.materials.xacro" />
 
   <!-- Import RRBot ros2_control description -->
-  <xacro:include filename="$(find rrbot_description)/ros2_control/rrbot.ros2_control.xacro" />
+  <xacro:include filename="$(find ros2_control_demo_example_1)/ros2_control/rrbot.ros2_control.xacro" />
 
   <!-- Used for fixing robot -->
   <link name="world"/>
