+++ example_1/hardware/include/ros2_control_demo_example_1/rrbot.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef EXAMPLE_1__HARDWARE__RRBOT_HPP_
-#define EXAMPLE_1__HARDWARE__RRBOT_HPP_
+#ifndef ROS2_CONTROL_DEMO_EXAMPLE_1__RRBOT_HPP_
+#define ROS2_CONTROL_DEMO_EXAMPLE_1__RRBOT_HPP_
 
 #include <memory>
 #include <string>
@@ -26,42 +26,42 @@
 #include "rclcpp/macros.hpp"
 #include "rclcpp_lifecycle/node_interfaces/lifecycle_node_interface.hpp"
 #include "rclcpp_lifecycle/state.hpp"
-#include "ros2_control_demo_hardware/visibility_control.h"
+#include "ros2_control_demo_example_1/visibility_control.h"
 
-namespace ros2_control_demo_hardware
+namespace ros2_control_demo_example_1
 {
 class RRBotSystemPositionOnlyHardware : public hardware_interface::SystemInterface
 {
 public:
   RCLCPP_SHARED_PTR_DEFINITIONS(RRBotSystemPositionOnlyHardware);
 
-  ROS2_CONTROL_DEMO_HARDWARE_PUBLIC
+  ROS2_CONTROL_DEMO_EXAMPLE_1_PUBLIC
   hardware_interface::CallbackReturn on_init(
     const hardware_interface::HardwareInfo & info) override;
 
-  ROS2_CONTROL_DEMO_HARDWARE_PUBLIC
+  ROS2_CONTROL_DEMO_EXAMPLE_1_PUBLIC
   hardware_interface::CallbackReturn on_configure(
     const rclcpp_lifecycle::State & previous_state) override;
 
-  ROS2_CONTROL_DEMO_HARDWARE_PUBLIC
+  ROS2_CONTROL_DEMO_EXAMPLE_1_PUBLIC
   std::vector<hardware_interface::StateInterface> export_state_interfaces() override;
 
-  ROS2_CONTROL_DEMO_HARDWARE_PUBLIC
+  ROS2_CONTROL_DEMO_EXAMPLE_1_PUBLIC
   std::vector<hardware_interface::CommandInterface> export_command_interfaces() override;
 
-  ROS2_CONTROL_DEMO_HARDWARE_PUBLIC
+  ROS2_CONTROL_DEMO_EXAMPLE_1_PUBLIC
   hardware_interface::CallbackReturn on_activate(
     const rclcpp_lifecycle::State & previous_state) override;
 
-  ROS2_CONTROL_DEMO_HARDWARE_PUBLIC
+  ROS2_CONTROL_DEMO_EXAMPLE_1_PUBLIC
   hardware_interface::CallbackReturn on_deactivate(
     const rclcpp_lifecycle::State & previous_state) override;
 
-  ROS2_CONTROL_DEMO_HARDWARE_PUBLIC
+  ROS2_CONTROL_DEMO_EXAMPLE_1_PUBLIC
   hardware_interface::return_type read(
     const rclcpp::Time & time, const rclcpp::Duration & period) override;
 
-  ROS2_CONTROL_DEMO_HARDWARE_PUBLIC
+  ROS2_CONTROL_DEMO_EXAMPLE_1_PUBLIC
   hardware_interface::return_type write(
     const rclcpp::Time & time, const rclcpp::Duration & period) override;
 
@@ -76,6 +76,6 @@ private:
   std::vector<double> hw_states_;
 };
 
-}  // namespace ros2_control_demo_hardware
+}  // namespace ros2_control_demo_example_1
 
-#endif  // EXAMPLE_1__HARDWARE__RRBOT_HPP_
+#endif  // ROS2_CONTROL_DEMO_EXAMPLE_1__RRBOT_HPP_
