+++ ros2_control_demo_hardware/ros2_control_demo_hardware.xml
@@ -1,11 +1,4 @@
 <library path="ros2_control_demo_hardware">
-  <class name="ros2_control_demo_hardware/RRBotSystemPositionOnlyHardware"
-         type="ros2_control_demo_hardware::RRBotSystemPositionOnlyHardware"
-         base_class_type="hardware_interface::SystemInterface">
-    <description>
-      The ROS2 Control minimal robot protocol. This is example for start porting the robot from ROS 1.
-    </description>
-  </class>
   <class name="ros2_control_demo_hardware/RRBotSystemMultiInterfaceHardware"
          type="ros2_control_demo_hardware::RRBotSystemMultiInterfaceHardware"
          base_class_type="hardware_interface::SystemInterface">
